requirements.txt
pandas==3.0.0
Pillow==12.1.0
scikit_learn==1.8.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
xgboost==3.1.3